requirements.txt
Django==5.0.7
gunicorn==22.0.0
psycopg[binary]==3.2.1
dj-database-url==2.2.0
whitenoise==6.7.0
pandas==2.2.2
numpy==1.26.4
python-dateutil==2.9.0.post0
